+++ forui_hooks/lib/src/tooltip_controller_hook.dart
@@ -8,13 +8,11 @@ import 'package:forui/forui.dart';
 FTooltipController useFTooltipController({
   TickerProvider? vsync,
   bool shown = false,
-  FTooltipMotion motion = const FTooltipMotion(),
   List<Object?>? keys,
 }) => use(
   _TooltipControllerHook(
     vsync: vsync ??= useSingleTickerProvider(keys: keys),
     shown: shown,
-    motion: motion,
     keys: keys,
   ),
 );
@@ -22,9 +20,8 @@ FTooltipController useFTooltipController({
 class _TooltipControllerHook extends Hook<FTooltipController> {
   final TickerProvider vsync;
   final bool shown;
-  final FTooltipMotion motion;
 
-  const _TooltipControllerHook({required this.vsync, required this.shown, required this.motion, super.keys});
+  const _TooltipControllerHook({required this.vsync, required this.shown, super.keys});
 
   @override
   _TooltipControllerHookState createState() => .new();
@@ -34,13 +31,12 @@ class _TooltipControllerHook extends Hook<FTooltipController> {
     super.debugFillProperties(properties);
     properties
       ..add(DiagnosticsProperty('vsync', vsync))
-      ..add(FlagProperty('shown', value: shown, ifTrue: 'shown'))
-      ..add(DiagnosticsProperty('motion', motion));
+      ..add(FlagProperty('shown', value: shown, ifTrue: 'shown'));
   }
 }
 
 class _TooltipControllerHookState extends HookState<FTooltipController, _TooltipControllerHook> {
-  late final _controller = FTooltipController(vsync: hook.vsync, shown: hook.shown, motion: hook.motion);
+  late final _controller = FTooltipController(vsync: hook.vsync, shown: hook.shown);
 
   @override
   FTooltipController build(BuildContext context) => _controller;